rename app-store to store
